+++ spec/Pim/Component/Catalog/Completeness/Checker/PriceCompleteCheckerSpec.php
@@ -1,30 +1,39 @@
 <?php
 
-namespace spec\Pim\Component\Catalog\Completeness\Checker\Attribute;
+namespace spec\Pim\Component\Catalog\Completeness\Checker;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
 use Pim\Bundle\CatalogBundle\Model\ChannelInterface;
 use Pim\Bundle\CatalogBundle\Model\CurrencyInterface;
+use Pim\Bundle\CatalogBundle\Model\LocaleInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductPriceInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductValueInterface;
 
 class PriceCompleteCheckerSpec extends ObjectBehavior
 {
+    public function it_is_a_completeness_checker()
+    {
+        $this->shouldImplement('Pim\Component\Catalog\Completeness\Checker\ProductValueCompleteCheckerInterface');
+    }
+
     public function it_suports_price_collection_attribute(
+        ProductValueInterface $productValue,
         AttributeInterface $attribute
     ) {
+        $productValue->getAttribute()->willReturn($attribute);
         $attribute->getAttributeType()->willReturn('pim_catalog_price_collection');
-        $this->supportsAttribute($attribute)->shouldReturn(true);
+        $this->supportsValue($productValue)->shouldReturn(true);
 
         $attribute->getAttributeType()->willReturn('other');
-        $this->supportsAttribute($attribute)->shouldReturn(false);
+        $this->supportsValue($productValue)->shouldReturn(false);
     }
 
     public function it_succesfully_checks_complete_price_collection(
         ProductValueInterface $value,
         ChannelInterface $channel,
+        LocaleInterface $locale,
         ArrayCollection $arrayCollection,
         CurrencyInterface $currency1,
         CurrencyInterface $currency2,
@@ -43,12 +52,13 @@ class PriceCompleteCheckerSpec extends ObjectBehavior
         $price2->getData()->willReturn(777);
 
         $value->getData()->willReturn([$price1, $price2]);
-        $this->isComplete($value, $channel, 'en_US')->shouldReturn(true);
+        $this->isComplete($value, $channel, $locale)->shouldReturn(true);
     }
 
     public function it_succesfully_checks_incomplete_price_collection(
         ProductValueInterface $value,
         ChannelInterface $channel,
+        LocaleInterface $locale,
         ArrayCollection $arrayCollection,
         CurrencyInterface $currency1,
         CurrencyInterface $currency2,
@@ -63,6 +73,6 @@ class PriceCompleteCheckerSpec extends ObjectBehavior
         $price1->getData()->willReturn(null);
 
         $value->getData()->willReturn([$price1]);
-        $this->isComplete($value, $channel, 'en_US')->shouldReturn(false);
+        $this->isComplete($value, $channel, $locale)->shouldReturn(false);
     }
 }